splits Tabs component into DatabaseTabs.vue and SideBarTabs.vue components, add tabsUtils mixin
